% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZ_bootstrap.R
\name{boot_nopodec_mean}
\alias{boot_nopodec_mean}
\title{Bootstrap of the mean decomposition procedure.}
\usage{
boot_nopodec_mean(data, treatment, variables, y, weights = NULL, R = 10,
  counterfactual = c("AB", "BA"), ...)
}
\arguments{
\item{data}{data frame}

\item{treatment}{column name of the binary variable}

\item{variables}{character vector of the variables' names, for which the common support has to be established}

\item{y}{name of the outcome variable for which you want to make the decomposition}

\item{weights}{name of the weight variable (sample weights). If \code{NULL}
(default value) it uses equal weights for all observations, adding
a column of ones}

\item{R}{The number of bootstrap replicates. This will be a single positive integer.}

\item{counterfactual}{"AB" or "BA". "AB" means that we want to estimate the counterfactual (wage) of group A, as if their characteristics were distributed as in group B. "BA" is the opposite (characteristics of group B are balanced to those of group A).}

\item{...}{other arguments passed on to \code{\link{boot}}.}
}
\description{
Bootstrap estimates of \code{\link{nopodec_mean}} and \code{\link{nopodec}}.
This serves to estimate for example confidence intervals of each component
of the decomposition.
}
\examples{
\dontrun{
data(invented_wages)
b01 <- boot_nopodec_mean(data = invented_wages, treatment = "gender",
                        variables = c("sector", "education"), y = "wage",
                        R = 500)
}

}
