% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_support.R
\name{common_support2}
\alias{common_support2}
\title{Establishes the Common Support between two groups of individuals, based on the (empirical) joint distribution of some characteristics (variables).}
\usage{
common_support2(data, treatment, variables)
}
\arguments{
\item{data}{data frame}

\item{treatment}{column name of the binary variable}

\item{variables}{character vector of the variables' names, for which the common support has to be established}
}
\value{
a data frame with all the combinations of the modalities of the variables considered. In addition two columns: \code{common_support} (logical, TRUE if in the common support), \code{strata} (identifier of a unique combination of the modalities of the variables)
}
\description{
Establishes the Common Support between two groups of individuals, based on the (empirical) joint distribution of some characteristics (variables).
}
\examples{
data(invented_wages)
common_support2(invented_wages, treatment = "gender",
                variables = c("sector", "education"))

}
\seealso{
See \code{\link{common_support_strata2}}, which uses this function and then joins with the original data.
}
