% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_support_strata.R
\name{common_support_strata2}
\alias{common_support_strata2}
\title{Establishes Common Support and joins with the data.}
\usage{
common_support_strata2(data, treatment, variables, y, weights = NULL)
}
\arguments{
\item{data}{data frame}

\item{treatment}{column name of the binary variable}

\item{variables}{character vector of the variables' names, for which the common support has to be established}

\item{y}{name of the outcome variable for which you want to make the decomposition}

\item{weights}{name of the weight variable (sample weights). If \code{NULL}
(default value) it uses equal weights for all observations, adding
a column of ones}
}
\value{
a data frame with the same number of rows of the starting data, with columns of the chosen treatment, variables, y, weights, common_support (TRUE/FALSE) and strata (a strata is given by a certain combination of the modalities of the variables)
}
\description{
Establishes Common Support and joins with the data.
}
\examples{
data(invented_wages)
common_support_strata2(invented_wages, treatment = "gender",
                       variables = c("sector", "education"),
                       y = "wage", weights = "sample_weights")

}
\seealso{
See \code{\link{common_support2}}, which is used by this function to establish common support.
    See \code{\link{fhat_strata2}} and \code{\link{reweight_strata2}}, that do the next steps.
    See \code{\link{reweight_strata_all2}} that does all the steps of \code{\link{common_support_strata2}},
     \code{\link{fhat_strata2}} and \code{\link{reweight_strata2}}.
}

