% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decr.R
\docType{package}
\name{decr}
\alias{decr}
\alias{decr-package}
\title{Tools to Perform Decompositions of Differences between two Groups}
\description{
Tools to perform decompositions of differences between two groups
(for example wage decompositions). Focus is on the common support
between the distribution of the characteristics of the two groups.
In the common support, counterfactuals can be estimated in order
to perform the decomposition.
}

