% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhat_strata.R
\name{fhat_strata2}
\alias{fhat_strata2}
\title{Estimates empirical joint frequencies of strata, in and out of the common support.}
\usage{
fhat_strata2(.cs_strata)
}
\arguments{
\item{.cs_strata}{the output of \code{\link{common_support_strata2}}}
}
\value{
a data frame with number of rows equal to the number of strata multiplied by 2 (for each group), with the following columns:
\itemize{
\item \code{strata} character vector with the combinations of the modalities of the considered variables;
\item \code{name} of the column used as the binary treatment;
\item \code{common_support} logical that indicates wether the strata is in or out the common support;
\item \code{Nhat} estimate of the number of individuals in the strata;
\item \code{fhat} estimate of the relative frequency of the strata for each group, in and out the common support;
\item \code{fhat_groups} estimate of the relative frequency of the strata for each group (without the distinction of in and out the common support).
}
}
\description{
Estimates empirical joint frequencies of strata, in and out of the common support.
}
\examples{
data(invented_wages)
c00 <- common_support_strata2(invented_wages, treatment = "gender",
                       variables = c("sector", "education"),
                       y = "wage", weights = "sample_weights")

fhat_strata2(c00)

}
\seealso{
\code{\link{common_support_strata2}}, whose result is the input of this function.
      See \code{\link{reweight_strata2}}, that takes the result of this function as its input.
}

