% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/various_dec.R
\name{margin_difference}
\alias{margin_difference}
\alias{margin_difference.default}
\title{Computes observed (marginal) difference between statistics of two groups (mean, quantiles, ...)}
\usage{
margin_difference(...)

\method{margin_difference}{default}(.margin_stat, ...)
}
\arguments{
\item{...}{arguments passed to or from other methods.}

\item{.margin_stat}{data frame with the marginal statistics of two groups. This is the output of function \code{\link{margin_mean}} or \code{\link{margin_quantile}}.}
}
\description{
Computes observed (marginal) difference between statistics of two groups (mean, quantiles, ...)
}
\examples{
data(invented_wages)

r00 <- reweight_strata_all2(invented_wages, treatment = "gender",
                            variables = c("sector", "education"),
                            y = "wage", weights = "sample_weights")

m00 <- margin_mean(r00)

m00

# Difference between observed average wages of men and women
margin_difference(m00)


# An example with quantiles. In this case, the 75th percentile
q00 <- margin_quantile(r00, probs = 0.75)

q00

# Difference between observed 75th percentiles of wages of men and women
margin_difference(q00)


}
