% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nopodec_Fhat.R
\name{margin_Fhat}
\alias{margin_Fhat}
\title{Empirical cumulative distribution function of y for groups A and B
(marginals) and the number of observations.}
\usage{
margin_Fhat(...)
}
\arguments{
\item{...}{arguments passed to or from other methods.}
}
\description{
Empirical cumulative distribution function of y for groups A and B
(marginals) and the number of observations.
}

