% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nopodec_mean.R
\name{margin_mean}
\alias{margin_mean}
\alias{margin_mean.default}
\alias{margin_mean.reweighted}
\title{Estimates average y for groups A and B (marginals) and the number of observations.}
\usage{
margin_mean(...)

\method{margin_mean}{default}(.reweight_strata_all, y = NULL,
  weights = NULL, ...)

\method{margin_mean}{reweighted}(.reweighted, ...)
}
\arguments{
\item{...}{arguments passed to or from other methods.}

\item{.reweight_strata_all}{output of \code{\link{reweight_strata_all2}}}

\item{y}{name of the outcome variable for which you want to make the decomposition. If NULL (default), the value is inherited from the attributes of \code{.reweight_strata_all}}

\item{weights}{name of the weight variable (sample weights). If NULL (default), the value is inherited from the attributes of \code{.reweight_strata_all}.}

\item{.reweighted}{an object of class \code{reweighted} (the output of \code{\link{reweight_strata_all4}})}
}
\value{
a data frame with two rows (one for each group) and the follwing three columns:
\itemize{
\item the name of the treatment column used in \code{\link{reweight_strata_all2}};
\item \code{ybar}: average of the y variable;
\item \code{Nhat}: estimate of the number of individuals.
}
}
\description{
Estimates average y for groups A and B (marginals) and the number of observations.
}
\examples{
data(invented_wages)

r00 <- reweight_strata_all2(invented_wages, treatment = "gender",
                            variables = c("sector", "education"),
                            y = "wage", weights = "sample_weights")

margin_mean(r00)

data(invented_wages)
r00 <- reweight_strata_all4(invented_wages, treatment = "gender",
                       variables = c("sector", "education"),
                       y = "wage", weights = "sample_weights")

str(r00)
names(r00)
class(r00)

margin_mean(r00)

}
