% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/various_dec.R
\name{margin_quantile}
\alias{margin_quantile}
\alias{margin_quantile.default}
\alias{margin_quantile.reweighted}
\title{Estimates quantiles (of level probs) of y for groups A and B (marginals) and the number of observations.}
\usage{
margin_quantile(...)

\method{margin_quantile}{default}(.reweight_strata_all, y = NULL,
  weights = NULL, probs = 0.5, ...)

\method{margin_quantile}{reweighted}(.reweighted, ...)
}
\arguments{
\item{...}{arguments passed to or from other methods.}

\item{.reweight_strata_all}{output of \code{\link{reweight_strata_all2}}}

\item{y}{name of the outcome variable for which you want to make the decomposition. If NULL (default), the value is inherited from the attributes of \code{.reweight_strata_all}}

\item{weights}{name of the weight variable (sample weights). If NULL (default), the value is inherited from the attributes of \code{.reweight_strata_all}.}

\item{probs}{numeric vector of length one with the desired quantile level (should be between 0 and 1).}

\item{.reweighted}{an object of class \code{reweighted} (the output of \code{\link{reweight_strata_all4}})}
}
\value{
a data frame with two rows (one for each group) and the follwing four columns:
\itemize{
\item the name of the treatment column used in \code{\link{reweight_strata_all2}};
\item \code{yhat}: quantile of the y variable;
\item \code{Nhat}: estimate of the number of individuals;
\item \code{probs}: level of the estimated quantile (between 0 and 1).
}
}
\description{
Estimates quantiles (of level probs) of y for groups A and B (marginals) and the number of observations.
}
\examples{
data(invented_wages)

r00 <- reweight_strata_all2(invented_wages, treatment = "gender",
                            variables = c("sector", "education"),
                            y = "wage", weights = "sample_weights")

margin_quantile(r00, probs = 0.75)

data(invented_wages)
r00 <- reweight_strata_all4(invented_wages, treatment = "gender",
                       variables = c("sector", "education"),
                       y = "wage", weights = "sample_weights")

str(r00)
names(r00)
class(r00)

margin_quantile(r00)

}
