<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nopodec.R
\name{nopodec}
\alias{nopodec}
\title{Decomposition of the observed difference between averages of y of group A and B in 4 components, as in Nopo (2008).}
\usage{
nopodec(.nopodec_, counterfactual = c("AB", "BA"))
}
\arguments{
\item{.nopodec_}{output of \code{\link{nopodec_mean}}.}

\item{counterfactual}{"AB" or "BA". "AB" means that we want to estimate the counterfactual (wage) of group A, as if their characteristics were distributed as in group B. "BA" is the opposite (characteristics of group B are balanced to those of group A).}
}
\value{
a list with five components:

\itemize{
  \item \code{delta_total}: total observed difference between average (wages)
   of group A and B;
  \item \code{delta_A}: part of the observed difference due to the fact that
  there are indidivuals of group A which are not comparable, for features,
   to individuals of group B;
  \item \code{delta_X} part explained by the fact that the two groups have a
   different distribution of characteristics;
  \item \code{delta_S} part not justified by the different distributions of the
   characteristics of the two groups, and potentially due to a difference
    in the remuneration structures between the two groups;
  \item \code{delta_B} part of the difference due to the fact that there are
   individuals of group B with characteristics that none of the group A
    has.
}
}
\description{
The decomposition of the difference is like that developed by Nopo (2008), especially
regarding the two components out of support. The decomposition in the common support
is instead done with the reweighting approach proposed by Di Nardo, Fortin and Lemieux (1996).
}
\examples{
data(invented_wages)

# Common support and computation of counterfactual weights
r00 <- reweight_strata_all2(data = invented_wages,
                            treatment = "gender",
                            variables = c("sector", "education"),
                            y = "wage",
                            weights = "sample_weights")

# Computation of the elements necessary to the decomposition
n00 <- nopodec_mean(r00)

# Nopo decomposition
nopodec(n00, counterfactual = "AB")

}
\references{
Nopo, H. 2008. Matching as a Tool to Decompose Wage Gaps. \emph{Review of Economics and Statistics}, 90(2): 290-299.

DiNardo J., N. M. Fortin and T. Lemieux. 1996. Labor Market Institutions and the Distribution of Wages, 1973-1992: A Semiparametric Approach. \emph{Econometrica} 64 (5): 1001-44
}
=======
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nopodec.R
\name{nopodec}
\alias{nopodec}
\title{Decomposition of the observed difference between averages of y of group A and B in 4 components, as in Nopo (2008).}
\usage{
nopodec(.nopodec_, counterfactual = c("AB", "BA"))
}
\arguments{
\item{.nopodec_}{output of \code{\link{nopodec_mean}}.}

\item{counterfactual}{"AB" or "BA". "AB" means that we want to estimate the counterfactual (wage) of group A, as if their characteristics were distributed as in group B. "BA" is the opposite (characteristics of group B are balanced to those of group A).}
}
\value{
a list with five components:

\itemize{
  \item \code{delta_total}: total observed difference between average (wages)
   of group A and B;
  \item \code{delta_A}: part of the observed difference due to the fact that
  there are indidivuals of group A which are not comparable, for features,
   to individuals of group B;
  \item \code{delta_X} part explained by the fact that the two groups have a
   different distribution of characteristics;
  \item \code{delta_S} part not justified by the different distributions of the
   characteristics of the two groups, and potentially due to a difference
    in the remuneration structures between the two groups;
  \item \code{delta_B} part of the difference due to the fact that there are
   individuals of group B with characteristics that none of the group A
    has.
}
}
\description{
The decomposition of the difference is like that developed by Nopo (2008), especially
regarding the two components out of support. The decomposition in the common support
is instead done with the reweighting approach proposed by Di Nardo, Fortin and Lemieux (1996).
}
\examples{
data(invented_wages)

# Common support and computation of counterfactual weights
r00 <- reweight_strata_all2(data = invented_wages,
                            treatment = "gender",
                            variables = c("sector", "education"),
                            y = "wage",
                            weights = "sample_weights")

# Computation of the elements necessary to the decomposition
n00 <- nopodec_mean(r00)

# Nopo decomposition
nopodec(n00, counterfactual = "AB")

}
\references{
Nopo, H. 2008. Matching as a Tool to Decompose Wage Gaps. \emph{Review of Economics and Statistics}, 90(2): 290-299.

DiNardo J., N. M. Fortin and T. Lemieux. 1996. Labor Market Institutions and the Distribution of Wages, 1973-1992: A Semiparametric Approach. \emph{Econometrica} 64 (5): 1001-44
}

>>>>>>> 872623d5fa9f931056dfaa972a10af7718e86ef4
