% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nopodec_Fhat.R
\name{nopodec_Fhat}
\alias{nopodec_Fhat}
\alias{nopodec_Fhat.default}
\alias{nopodec_Fhat.reweighted}
\title{Empirical cumulative distribution function (ecdf) at value y for the two groups, in and out the common support.
In the common support, counterfactual ecdf are estimated.
It also estimates the number of individuals of the two groups.}
\usage{
nopodec_Fhat(...)

\method{nopodec_Fhat}{default}(.reweight_strata_all, y = NULL,
  weights = NULL, value, ...)

\method{nopodec_Fhat}{reweighted}(.reweighted, value, ...)
}
\arguments{
\item{...}{arguments passed to or from other methods.}

\item{.reweight_strata_all}{output of \code{\link{reweight_strata_all2}}}

\item{y}{name of the outcome variable for which you want to make the decomposition. If NULL (default), the value is inherited from the attributes of \code{.reweight_strata_all}}

\item{weights}{name of the weight variable (sample weights). If NULL (default), the value is inherited from the attributes of \code{.reweight_strata_all}.}

\item{value}{value of y for which the ecdf is estimated}

\item{.reweighted}{an object of class \code{reweighted} (the output of \code{\link{reweight_strata_all4}})}
}
\description{
The results are all the components necessary to perform a
decomposition of the ecdf difference between two groups,
in four components (as in Nopo (2004)).
}
\examples{
data(invented_wages)
r00 <- reweight_strata_all2(invented_wages, treatment = "gender",
                       variables = c("sector", "education"),
                       y = "wage", weights = "sample_weights")

nopodec_Fhat(r00, value = 5000)

r00 <- reweight_strata_all4(invented_wages, treatment = "gender",
                       variables = c("sector", "education"),
                       y = "wage", weights = "sample_weights")

str(r00)
names(r00)
class(r00)

nopodec_Fhat(r00, value = 5000)

}
