% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nopodec_mean.R
\name{nopodec_mean}
\alias{nopodec_mean}
\alias{nopodec_mean.default}
\alias{nopodec_mean.reweighted}
\title{Estimates y averages for the two groups, in and out the common support.
In the common support, counterfactual average y are estimated.
It also estimates the number of individuals of the two groups.}
\usage{
nopodec_mean(...)

\method{nopodec_mean}{default}(.reweight_strata_all, y = NULL,
  weights = NULL, ...)

\method{nopodec_mean}{reweighted}(.reweighted, ...)
}
\arguments{
\item{...}{arguments passed to or from other methods.}

\item{.reweight_strata_all}{output of \code{\link{reweight_strata_all2}}}

\item{y}{name of the outcome variable for which you want to make the decomposition. If NULL (default), the value is inherited from the attributes of \code{.reweight_strata_all}}

\item{weights}{name of the weight variable (sample weights). If NULL (default), the value is inherited from the attributes of \code{.reweight_strata_all}.}

\item{.reweighted}{an object of class \code{reweighted} (the output of \code{\link{reweight_strata_all4}})}
}
\value{
A data frame with two, three or four rows, with the following columns:
\itemize{
\item the name of the treatment column used in \code{\link{reweight_strata_all2}};
\item \code{common_support} logical indicating if in or out the common support;
\item \code{ybar} average of the y variable, weighted by the given weights;
\item \code{ybar_C_A} counterfactual average y of group A as if they had the same distribution of characteristics of group B. This is computed in the common support only and for group A individuals. It is computed with the weights \code{w_AB} that result from \code{\link{reweight_strata_all2}};
\item \code{ybar_C_B} counterfactual average y of group B as if they had the same distribution of characteristics of group A. This is computed in the common support only and for group B individuals. It is computed with the weights \code{w_BA} that result from \code{\link{reweight_strata_all2}};
\item \code{Nhat} estimate of the number of individuals.
}
The number of rows is given by the combinations of the distinct values of
the first two columns: \code{treatment} and \code{common_support}.
In the "typical" case, the resulting data frame will have 4 rows. It can have three rows if all the individuals of one group are in the common support.
In case of no common support or no out-of-support, the data frame will have two rows.
}
\description{
The results are all the components necessary to perform a
decomposition of the average wage difference between two groups,
in four components (as in Nopo (2004)).
}
\examples{
data(invented_wages)
r00 <- reweight_strata_all2(invented_wages, treatment = "gender",
                       variables = c("sector", "education"),
                       y = "wage", weights = "sample_weights")

nopodec_mean(r00)

data(invented_wages)
r00 <- reweight_strata_all4(invented_wages, treatment = "gender",
                       variables = c("sector", "education"),
                       y = "wage", weights = "sample_weights")

str(r00)
names(r00)
class(r00)

nopodec_mean(r00)

}

