% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reweight_strata.R
\name{reweight_strata2}
\alias{reweight_strata2}
\title{Computes the adjustment factors to estimate counterfactuals (in the common support).}
\usage{
reweight_strata2(.fhat_strata)
}
\arguments{
\item{.fhat_strata}{output of \code{\link{fhat_strata2}}}
}
\value{
a data frame with the \code{strata} in the common support (in the rows), the reweighting factors \code{rw_BA} and \code{rw_AB} (see below), and joint frequencies of the strata in the two groups (\code{f_A} and \code{f_B}).

Details about the columns of the returned data frame:
\item{strata}{Unique combinations of the modalities of the variables used in \code{\link{common_support_strata2}} (and \code{\link{common_support2}}). In this function, the strata are only those on the common support (that is, where there exists at least one observation of each group - A and B).}
\item{rw_BA}{This is the reweighting factor which has to be given to group B
 such that the distribution of their characteristics will be the same as that
  of group A. This is computed, in each \code{strata}, as \code{f_A / f_B}.}
\item{rw_AB}{This is the reweighting factor which has to be given to group A such that the distribution of their characteristics will be the same as that of group B. This is computed, in each \code{strata}, as \code{f_B / f_A}.}
\item{f_A}{estimated frequencies of \code{strata} for group A (in the common support).}
\item{f_B}{estimated frequencies of \code{strata} for group B (in the common support).}
}
\description{
Adjustment factors are used to reweight the joint distribution of characteristics of
one group so that it will be equal to that of the other group
(both counterfacuals, AB and BA...). This will be done on the
common support only.
}
\examples{
data(invented_wages)
c00 <- common_support_strata2(invented_wages, treatment = "gender",
                       variables = c("sector", "education"),
                       y = "wage", weights = "sample_weights")

f00 <- fhat_strata2(c00)
reweight_strata2(f00)

}
\seealso{
\code{\link{fhat_strata2}}, whose result is the input of this function.
       See \code{\link{reweight_strata_all2}}, that uses internally functions
       \code{\link{common_support_strata2}}, \code{\link{fhat_strata2}} and \code{\link{reweight_strata2}}.
}
