% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reweight_strata_all.R
\name{reweight_strata_all2}
\alias{reweight_strata_all2}
\title{Estimates adjustment (counterfactual) weights in the common support and joins with the data.
For observations out of the common support, the weights are equal to the sample weights (or one, in case of absence of sample weights).}
\usage{
reweight_strata_all2(data, treatment, variables, y, weights = NULL)
}
\arguments{
\item{data}{data frame}

\item{treatment}{column name of the binary variable}

\item{variables}{character vector of the variables' names, for which the common support has to be established}

\item{y}{name of the outcome variable for which you want to make the decomposition}

\item{weights}{name of the weight variable (sample weights). If \code{NULL}
(default value) it uses equal weights for all observations, adding
a column of ones}
}
\value{
a data frame with the same number of rows of the starting data, with columns of the chosen \code{treatment}, \code{variables}, \code{y}, \code{weights}, \code{common_support} (\code{TRUE} or \code{FALSE}), \code{strata} (a \code{strata} is given by a certain combination of the modalities of the \code{variables}), adjustment factors (\code{rw_BA} and \code{rw_AB}), empirical joint frequencies of strata in the two groups (\code{f_A} and \code{f_B}) and adjustment weights (\code{w_BA} and \code{w_AB}), which have to be used to perform (wage) decompositions
}
\description{
Estimates adjustment (counterfactual) weights in the common support and joins with the data.
For observations out of the common support, the weights are equal to the sample weights (or one, in case of absence of sample weights).
}
\examples{
data(invented_wages)
reweight_strata_all2(invented_wages, treatment = "gender",
                       variables = c("sector", "education"),
                       y = "wage", weights = "sample_weights")

}
\seealso{
\code{\link{common_support_strata2}}, \code{\link{fhat_strata2}} and
         \code{\link{reweight_strata2}}, that are used internally by this function.
         For the next steps, see for example \code{\link{nopodec_mean}}, \code{\link{dec_median}}
         and \code{\link{dec_quantile}}.
         See also \code{\link{margin_mean}} and \code{\link{margin_quantile}}, that use the output
         of this function as their input.
         This function is also used in \code{\link{reweight_strata_all4}}, whose result is an object of class
         \code{reweighted}, that is a list with 3 components: the output of \code{\link{fhat_strata2}}
         (\code{.fhat_strata}), \code{\link{reweight_strata2}} (\code{.reweight_strata}) and
         \code{\link{reweight_strata_all2}} (\code{.reweight_strata_all}).
}
