% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/various_dec.R
\name{dec_}
\alias{dec_}
\title{Performs decomposition of the quantile difference (of y) between two groups (in 2 components) in the common support.}
\usage{
dec_(.dec_, counterfactual = c("AB", "BA"))
}
\arguments{
\item{.dec_}{output of \code{\link{dec_median}} or \code{\link{dec_quantile}}.}

\item{counterfactual}{"AB" or "BA". "AB" means that we want to estimate the counterfactual (wage) of group A, as if their characteristics were distributed as in group B. "BA" is the opposite (characteristics of group B are balanced to those of group A).}
}
\value{
a list with four components:

\itemize{
  \item \code{probs}: the chosen quantile level;
  \item \code{delta_tot}: total observed difference between the quantiles (of wages)
   of group A and B (all the sample);
  \item \code{delta_tot_CS}: total observed difference between the quantiles (of wages)
   of group A and B in the common support;
  \item \code{delta_AB}: difference explained by the fact that the two groups
   have some combinations of characteristics that the other group has not.
  \item \code{delta_X}: part explained by the fact that the two groups have a
   different distribution of characteristics (same combinations of characteristics)
   but distributed differently);
  \item \code{delta_S}: part not justified by the different distributions of the
   characteristics of the two groups, and potentially due to a difference
    in the remuneration structures between the two groups.
}
}
\description{
Performs decomposition of the quantile difference (of y) between two groups (in 2 components) in the common support.
}
\examples{
data(invented_wages)

# Common support and computation of counterfactual weights
r00 <- reweight_strata_all2(invented_wages, treatment = "gender",
                       variables = c("sector", "education"),
                       y = "wage", weights = "sample_weights")

# Computation of the elements necessary to the decomposition
d00 <- dec_quantile(r00)

# Decomposition of the difference of the medians of the two groups
# (in the common support) in two components
dec_(d00)

}
